+++ org-roam.el
@@ -1,12 +1,12 @@
 ;;; org-roam.el --- A database abstraction layer for Org-mode -*- coding: utf-8; lexical-binding: t; -*-
 
-;; Copyright © 2020-2021 Jethro Kuan <jethrokuan95@gmail.com>
+;; Copyright © 2020-2022 Jethro Kuan <jethrokuan95@gmail.com>
 
 ;; Author: Jethro Kuan <jethrokuan95@gmail.com>
 ;; URL: https://github.com/org-roam/org-roam
 ;; Keywords: org-mode, roam, convenience
-;; Version: 2.2.0
-;; Package-Requires: ((emacs "26.1") (dash "2.13") (org "9.4") (emacsql "3.0.0") (emacsql-sqlite "1.0.0") (magit-section "3.0.0"))
+;; Version: 2.2.2
+;; Package-Requires: ((emacs "26.1") (dash "2.13") (org "9.4") (emacsql "20230228") (magit-section "3.0.0"))
 
 ;; This file is NOT part of GNU Emacs.
 
@@ -83,6 +83,7 @@
 (require 'emacsql-sqlite)
 
 (require 'org)
+(require 'org-attach)                   ; To set `org-attach-id-dir'
 (require 'org-id)
 (require 'ol)
 (require 'org-element)
@@ -122,6 +123,12 @@ All Org files, at any level of nesting, are considered part of the Org-roam."
   :group 'org-roam
   :type 'hook)
 
+(defcustom org-roam-post-node-insert-hook nil
+  "Hook run when an Org-roam node is inserted as an Org link.
+Each function takes two arguments: the id of the node, and the link description."
+  :group 'org-roam
+  :type 'hook)
+
 (defcustom org-roam-file-extensions '("org")
   "List of file extensions to be included by Org-Roam.
 While a file extension different from \".org\" may be used, the
@@ -130,9 +137,11 @@ responsibility to ensure that."
   :type '(repeat string)
   :group 'org-roam)
 
-(defcustom org-roam-file-exclude-regexp nil
+(defcustom org-roam-file-exclude-regexp (list org-attach-id-dir)
   "Files matching this regular expression are excluded from the Org-roam."
   :type '(choice
+          (repeat
+           (string :tag "Regular expression matching files to ignore"))
           (string :tag "Regular expression matching files to ignore")
           (const :tag "Include everything" nil))
   :group 'org-roam)
@@ -143,31 +152,40 @@ responsibility to ensure that."
     '(find fd fdfind rg))
   "Commands that will be used to find Org-roam files.
 
-It should be a list of symbols or cons cells representing any of the following
-supported file search methods.
+It should be a list of symbols or cons cells representing any of
+the following supported file search methods.
 
-The commands will be tried in order until an executable for a command is found.
-The Elisp implementation is used if no command in the list is found.
+The commands will be tried in order until an executable for a
+command is found. The Elisp implementation is used if no command
+in the list is found.
 
   `find'
+
     Use find as the file search method.
     Example command:
-    find /path/to/dir -type f \( -name \"*.org\" -o -name \"*.org.gpg\" \)
+      find /path/to/dir -type f \
+        \( -name \"*.org\" -o -name \"*.org.gpg\" -name \"*.org.age\" \)
 
   `fd'
+
     Use fd as the file search method.
-    Example command: fd /path/to/dir/ --type file -e \".org\" -e \".org.gpg\"
+    Example command:
+      fd /path/to/dir/ --type file -e \".org\" -e \".org.gpg\" -e \".org.age\"
 
   `fdfind'
+
     Same as `fd'. It's an alias that used in some OSes (e.g. Debian, Ubuntu)
 
   `rg'
-    Use ripgrep as the file search method.
-    Example command: rg /path/to/dir/ --files -g \"*.org\" -g \"*.org.gpg\"
 
-By default, `executable-find' will be used to look up the path to the
-executable. If a custom path is required, it can be specified together with the
-method symbol as a cons cell. For example: '(find (rg . \"/path/to/rg\"))."
+    Use ripgrep as the file search method.
+    Example command:
+       rg /path/to/dir/ --files -g \"*.org\" -g \"*.org.gpg\" -g \"*.org.age\"
+
+By default, `executable-find' will be used to look up the path to
+the executable. If a custom path is required, it can be specified
+together with the method symbol as a cons cell. For example:
+'(find (rg . \"/path/to/rg\"))."
   :type '(set (const :tag "find" find)
               (const :tag "fd" fd)
               (const :tag "fdfind" fdfind)
@@ -183,19 +201,34 @@ FILE is an Org-roam file if:
 - It's located somewhere under `org-roam-directory'
 - It has a matching file extension (`org-roam-file-extensions')
 - It doesn't match excluded regexp (`org-roam-file-exclude-regexp')"
-  (let* ((path (or file (buffer-file-name (buffer-base-buffer))))
-         (ext (when path (org-roam--file-name-extension path)))
-         (ext (if (string= ext "gpg")
-                  (org-roam--file-name-extension (file-name-sans-extension path))
-                ext)))
-    (save-match-data
-      (and
-       path
-       (member ext org-roam-file-extensions)
-       (not (and org-roam-file-exclude-regexp
-                 (string-match-p org-roam-file-exclude-regexp path)))
-       (org-roam-descendant-of-p path (expand-file-name org-roam-directory))))))
+  (when (or file (buffer-file-name (buffer-base-buffer)))
+    (let* ((path (or file (buffer-file-name (buffer-base-buffer))))
+           (relative-path (file-relative-name path org-roam-directory))
+           (ext (org-roam--file-name-extension path))
+           (ext (if (or (string= ext "gpg")
+                        (string= ext "age"))
+                    (org-roam--file-name-extension (file-name-sans-extension path))
+                  ext))
+           (org-roam-dir-p (org-roam-descendant-of-p path org-roam-directory))
+           (valid-file-ext-p (member ext org-roam-file-extensions))
+           (match-exclude-regexp-p
+            (cond
+             ((not org-roam-file-exclude-regexp) nil)
+             ((stringp org-roam-file-exclude-regexp)
+              (string-match-p org-roam-file-exclude-regexp relative-path))
+             ((listp org-roam-file-exclude-regexp)
+              (let (is-match)
+                (dolist (exclude-re org-roam-file-exclude-regexp)
+                  (setq is-match (or is-match (string-match-p exclude-re relative-path))))
+                is-match)))))
+      (save-match-data
+        (and
+         path
+         org-roam-dir-p
+         valid-file-ext-p
+         (not match-exclude-regexp-p))))))
 
+;;;###autoload
 (defun org-roam-list-files ()
   "Return a list of all Org-roam files under `org-roam-directory'.
 See `org-roam-file-p' for how each file is determined to be as
@@ -260,14 +293,15 @@ If no files are found, an empty list is returned."
        (ansi-color-filter-apply it)
        (split-string it "\n")
        (seq-filter (lambda (s)
-                     (or (null s) (string= "" s))) it)))
+                     (not (or (null s) (string= "" s)))) it)))
 
 (defun org-roam--list-files-search-globs (exts)
   "Given EXTS, return a list of search globs.
 E.g. (\".org\") => (\"*.org\" \"*.org.gpg\")"
   (cl-loop for e in exts
            append (list (format "\"*.%s\"" e)
-                        (format "\"*.%s.gpg\"" e))))
+                        (format "\"*.%s.gpg\"" e)
+                        (format "\"*.%s.age\"" e))))
 
 (defun org-roam--list-files-find (executable dir)
   "Return all Org-roam files under DIR, using \"find\", provided as EXECUTABLE."
@@ -298,7 +332,7 @@ E.g. (\".org\") => (\"*.org\" \"*.org.gpg\")"
   "Return all Org-roam files under DIR, using Elisp based implementation."
   (let ((regex (concat "\\.\\(?:"(mapconcat
                                   #'regexp-quote org-roam-file-extensions
-                                  "\\|" )"\\)\\(?:\\.gpg\\)?\\'"))
+                                  "\\|" )"\\)\\(?:\\.gpg\\|\\.age\\)?\\'"))
         result)
     (dolist (file (org-roam--directory-files-recursively dir regex nil nil t) result)
       (when (and (file-readable-p file)
@@ -316,6 +350,7 @@ E.g. (\".org\") => (\"*.org\" \"*.org.gpg\")"
   (require 'org-roam-id)
   (require 'org-roam-capture)
   (require 'org-roam-mode)
+  (require 'org-roam-log)
   (require 'org-roam-migrate))
 
 ;;; org-roam.el ends here